Use minima remote theme.
The default minima theme has not been updated in two years, but many
features have been added to the GitHub repo of the minima theme, such as
color schemes and support for custom HTML heads.

This commit makes the website use the latest version of the minima
theme to take advantage of these features.