Add favicon to all pages.
Using the latest version of the minima theme allows creating a
`custom-head.html` file. This file is automatically added to every page
by Jekyll. I created this file and inserted links to the favicon images.
I also inserted a `site.manifest` file, which is useful for progressive
web applications. The website is not a WPA but it does not hurt to have
it.