Show first paragraph from post.
Without this option enabled, users can only see the post name and must
click on the post to see more information about the post.

With this option enabled, posts are listed along with the first
paragraph or the first image in the post. In this way users can see
what the post is about.